% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceblR-package.R
\docType{package}
\name{ceblR-package}
\alias{ceblR}
\alias{ceblR-package}
\title{ceblR: Extract Data from the Canadian Elite Basketball League}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Gather boxscore and play-by-play data from the CEBL website to create a repository of basic and advanced statistics for teams and players.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/awosoga/ceblR}
  \item \url{https://awosoga.github.io/ceblR/}
  \item Report bugs at \url{https://github.com/awosoga/ceblR/issues}
}

}
\author{
\strong{Maintainer}: David Awosoga \email{odo.awosoga@gmail.com} (\href{https://orcid.org/0000-0002-2741-5770}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Matthew Chow \email{mc2chow@uwaterloo.ca}
}

}
\keyword{internal}
