% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Remove spaces, tabs, and newlines from a string}
\usage{
clean_data(dirty_data)
}
\arguments{
\item{dirty_data}{A character string}
}
\value{
A character string  without spaces, tabs, and newline characters
}
\description{
Remove spaces, tabs, and newlines from a string
}
\examples{
dirty_data <- " \n CEBL "
clean_data(dirty_data)
}
