% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{load_cebl_officials}
\alias{load_cebl_officials}
\title{load_cebl_officials}
\usage{
load_cebl_officials(seasons = NULL)
}
\arguments{
\item{seasons}{A vector of seasons (years) to filter the data.}
}
\value{
A data frame containing the CEBL officials for the specified seasons.
}
\description{
A function to load the Canadian Elite Basketball League (CEBL) officials.
}
\examples{
load_cebl_officials(2020:2021)
================================    ===========
|Column Name                      |   Type    |
================================    ===========
|game_id                          |   dbl     |
|season                           |   dbl     |
|officials_type                   |   chr     |
|officials_name                   |   chr     |
|first_name                       |   chr     |
|last_name                        |   chr     |
|scoreboard_name                  |   chr     |
|first_name_initial               |   chr     |
|last_name_initial                |   chr     |
|international_first_name         |   chr     |
|international_first_name_initial |   chr     |
|international_last_name          |   chr     |
|international_last_name_initial  |   chr     |
|scoreboard_name                  |   chr     |
|================================   ===========

}
