% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{load_cebl_pbp}
\alias{load_cebl_pbp}
\title{load_cebl_pbp}
\usage{
load_cebl_pbp(seasons = NULL)
}
\arguments{
\item{seasons}{A vector of seasons (years) to filter the data.}
}
\value{
A data frame containing the CEBL play-by-play data for the specified seasons.
}
\description{
A function to load the Canadian Elite Basketball League (CEBL) play-by-play data.
}
\examples{
load_cebl_pbp(2020:2021)

================================   ===========
|Column Name                     |   Type   |
================================   ===========
|game_id                         |   dbl    |
|season                          |   dbl    |
|game_time                       |   chr    |
|home_score                      |   dbl    |
|away_score                      |   dbl    |
|home_lead                       |   dbl    |
|team_id                         |   dbl    |
|period                          |   dbl    |
|period_type                     |   chr    |
|player_id                       |   dbl    |
|scoreboard_name                 |   chr    |
|success                         |   dbl    |
|action_type                     |   chr    |
|action_number                   |   dbl    |
|previous_action                 |   dbl    |
|sub_type                        |   chr    |
|scoring                         |   dbl    |
|shirt_number                    |   dbl    |
|player_name                     |   chr    |
|first_name                      |   chr    |
|last_name                       |   chr    |
|x                               |   dbl    |
|y                               |   dbl    |
|qualifier_0                     |   chr    |
|qualifier_1                     |   chr    |
|qualifier_2                     |   chr    |
|qualifier_3                     |   chr    |
|international_first_name        |   chr    |
|international_last_name         |   chr    |
|international_first_name_initial|   chr    |
|international_last_name_initial |   chr    |
================================   ===========
}
