% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{load_cebl_schedule}
\alias{load_cebl_schedule}
\title{load_cebl_schedule}
\usage{
load_cebl_schedule(seasons = NULL)
}
\arguments{
\item{seasons}{A vector of seasons (years) to filter the data.
if NULL, defaults to the range from 2019 to the current year.}
}
\value{
A data frame containing the CEBL schedule for the specified seasons.
}
\description{
A function to load the Canadian Elite Basketball League (CEBL) schedule.
}
\examples{
load_cebl_schedule(2020:2021)

==============================================
|          Column Name           |   Type    |
==============================================
|fiba_id                         |   dbl     |
|season                          |   dbl     |
|start_time_utc                  |   dttm    |
|status                          |   chr     |
|competition                     |   chr     |
|venue_name                      |   chr     |
|period                          |   dbl     |
|home_team_id                    |   dbl     |
|home_team_name                  |   chr     |
|home_team_score                 |   dbl     |
|home_team_logo_url              |   chr     |
|home_team_url_stats_en          |   chr     |
|home_team_url_stats_fr          |   chr     |
|away_team_id                    |   dbl     |
|away_team_name                  |   chr     |
|away_team_score                 |   dbl     |
|away_team_logo_url              |   chr     |
|away_team_url_stats_en          |   chr     |
|away_team_url_stats_fr          |   chr     |
|stats_url_en                    |   chr     |
|stats_url_fr                    |   chr     |
|cebl_stats_url_en               |   chr     |
|cebl_stats_url_fr               |   chr     |
|tickets_url_en                  |   chr     |
|tickets_url_fr                  |   chr     |
|id                              |   dbl     |
|fiba_json_url                   |   chr     |
================================  ===========|

}
