% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{player_advanced_data}
\alias{player_advanced_data}
\title{Get CEBL player advanced data}
\usage{
player_advanced_data(year)
}
\arguments{
\item{year}{The season(s) of interest.}
}
\value{
A dataframe of player advanced data within the requested season(s).
}
\description{
Get CEBL player advanced data
}
\examples{
try({
player_advanced_data <- player_advanced_data(2020)
player_advanced_data_2019_and_2020 <- player_advanced_data(c(2019:2020))
})
}
