% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{player_boxscore_data}
\alias{player_boxscore_data}
\title{Get CEBL player boxscore data}
\usage{
player_boxscore_data(year)
}
\arguments{
\item{year}{The season(s) of interest.}
}
\value{
A dataframe of player boxscore data within the requested season(s).
}
\description{
Get CEBL player boxscore data
}
\examples{
try({
player_boxscores <- player_boxscore_data(2020)
player_boxscores_2019_and_2020 <- player_boxscore_data(c(2019:2020))
})
}
