% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{team_advanced_data}
\alias{team_advanced_data}
\title{Get CEBL team advanced data}
\usage{
team_advanced_data(year)
}
\arguments{
\item{year}{The season(s) of interest.}
}
\value{
A dataframe of team advanced data within the requested season(s).
}
\description{
Get CEBL team advanced data
}
\examples{
try({
team_advanced_data <- team_advanced_data(2020)
team_advanced_data_2019_and_2020 <- team_advanced_data(c(2019:2020))
})
}
