% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{team_boxscore_data}
\alias{team_boxscore_data}
\title{Get CEBL team boxscore data}
\usage{
team_boxscore_data(year)
}
\arguments{
\item{year}{The season(s) of interest.}
}
\value{
A dataframe of team boxscore data within the requested season(s).
}
\description{
Get CEBL team boxscore data
}
\examples{
try({
team_boxscores <- team_boxscore_data(2020)
team_boxscores_2019_and_2020 <- team_boxscore_data(c(2019:2020))
})
}
