% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_seasons}
\alias{validate_seasons}
\title{Checks whether all the provided seasons are valid years}
\usage{
validate_seasons(seasons)
}
\arguments{
\item{seasons}{A vector of seasons (years) to filter the data.
if NULL, defaults to the range from 2019 to the current year.}
}
\value{
A logical value indicating whether the seasons are valid.
}
\description{
Checks whether all the provided seasons are valid years
}
\examples{
validate_seasons(2020:2021)
}
