% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{load_cebl_coaches}
\alias{load_cebl_coaches}
\title{load_cebl_coaches}
\usage{
load_cebl_coaches(seasons = NULL)
}
\arguments{
\item{seasons}{A vector of seasons (years) to filter the data.}
}
\value{
A data frame containing the CEBL coaches for the specified seasons.
}
\description{
A function to load the Canadian Elite Basketball League (CEBL) coaches.
}
\examples{
load_cebl_coaches(2020:2021)

================================   ===========
}
