% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{load_cebl_officials}
\alias{load_cebl_officials}
\title{\strong{Load CEBL Officials}}
\usage{
load_cebl_officials(seasons = NULL)
}
\arguments{
\item{seasons}{A vector of seasons (years) to filter the data.}
}
\value{
A data frame containing the CEBL officials for the specified seasons.\tabular{ll}{
   Column Name \tab Type \cr
   game_id \tab dbl \cr
   season \tab dbl \cr
   officials_type \tab chr \cr
   officials_name \tab chr \cr
   first_name \tab chr \cr
   last_name \tab chr \cr
   scoreboard_name \tab chr \cr
   first_name_initial \tab chr \cr
   last_name_initial \tab chr \cr
   international_first_name \tab chr \cr
   international_first_name_initial \tab chr \cr
   international_last_name \tab chr \cr
   international_last_name_initial \tab chr \cr
   scoreboard_name \tab chr \cr
}
}
\description{
A function to load the Canadian Elite Basketball League (CEBL) officials.
}
\examples{
load_cebl_officials(2020:2021)

}
