% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{load_cebl_player_boxscores}
\alias{load_cebl_player_boxscores}
\title{load_cebl_player_boxscores}
\usage{
load_cebl_player_boxscores(seasons = NULL)
}
\arguments{
\item{seasons}{A vector of seasons (years) to filter the data.}
}
\value{
A data frame containing the CEBL player boxscores for the specified seasons.
}
\description{
A function to load the Canadian Elite Basketball League (CEBL) player boxscores.
}
\examples{
load_cebl_player_boxscores(2020:2021)

======================================   ===========
| Column Name                          |    Type   |
======================================   ===========
|game_id                               |   dbl    |
|season                                |   dbl    |
|team_name                             |   chr    |
|player_number                         |   dbl    |
|player_name                           |   chr    |
|player_position                       |   chr    |
|minutes                               |   chr    |
|podbls                                |   dbl    |
|field_goals_made                      |   dbl    |
|field_goals_attempted                 |   dbl    |
|field_goal_percentage                 |   dbl    |
|two_podbl_field_goals_made            |   dbl    |
|two_podbl_field_goals_attempted       |   dbl    |
|two_podbl__percentage                 |   dbl    |
|three_podbl_field_goals_made          |   dbl    |
|three_podbl_field_goals_attempted     |   dbl    |
|three_podbl_percentage                |   dbl    |
|free_throws_made                      |   dbl    |
|free_throws_attempted                 |   dbl    |
|free_throw_percentage                 |   dbl    |
|offensive_rebounds                    |   dbl    |
|defensive_rebounds                    |   dbl    |
|rebounds                              |   dbl    |
|assists                               |   dbl    |
|turnovers                             |   dbl    |
|steals                                |   dbl    |
|blocks                                |   dbl    |
|blocks_received                       |   dbl    |
|personal_fouls                        |   dbl    |
|fouls_drawn                           |   dbl    |
|plus_minus                            |   dbl    |
|index_rating                          |   dbl    |
|index_rating_2                        |   dbl    |
|index_rating_3                        |   dbl    |
|index_rating_4                        |   dbl    |
|index_rating_5                        |   dbl    |
|index_rating_6                        |   dbl    |
|index_rating_7                        |   lgl    |
|second_chance_points                  |   lgl    |
|fast_break_points                     |   lgl    |
|points_in_the_paint                   |   dbl    |
|first_name                            |   chr    |
|first_name_initial                    |   chr    |
|last_name                             |   chr    |
|last_name_initial                     |   chr    |
|international_first_name              |   chr    |
|international_first_name_initial      |   chr    |
|international_last_name               |   chr    |
|international_last_name_initial       |   chr    |
|scoreboard_name                       |   chr    |
|active                                |   lgl    |
|starter                               |   lgl    |
|captain                               |   lgl    |
|photo_t                               |   chr    |
|photo_s                               |   chr    |
======================================  ===========
}
