% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{load_cebl_team_boxscores}
\alias{load_cebl_team_boxscores}
\title{load_cebl_team_boxscores}
\usage{
load_cebl_team_boxscores(seasons = NULL)
}
\arguments{
\item{seasons}{A vector of seasons (years) to filter the data.}
}
\value{
A data frame containing the CEBL team boxscores for the specified seasons.
}
\description{
A function to load the Canadian Elite Basketball League (CEBL) team boxscores.
}
\examples{
load_cebl_team_boxscores(2020:2021)


===========================================================
|Column Name                                   |   Type   |
===========================================================
|game_id                                       |   dbl    |
|season                                        |   dbl    |
|team_name                                     |   chr    |
|short_name                                    |   chr    |
|code                                          |   chr    |
|team_score                                    |   dbl    |
|minutes                                       |   chr    |
|field_goals_made                              |   dbl    |
|field_goals_attempted                         |   dbl    |
|field_goal_percentage                         |   dbl    |
|two_point_field_goals_made                    |   dbl    |
|two_point_field_goals_attempted               |   dbl    |
|two_point__percentage                         |   dbl    |
|three_point_field_goals_made                  |   dbl    |
|three_point_field_goals_attempted             |   dbl    |
|three_point_percentage                        |   dbl    |
|free_throws_made                              |   dbl    |
|free_throws_attempted                         |   dbl    |
|free_throw_percentage                         |   dbl    |
|offensive_rebounds                            |   dbl    |
|defensive_rebounds                            |   dbl    |
|rebounds                                      |   dbl    |
|assists                                       |   dbl    |
|steals                                        |   dbl    |
|turnovers                                     |   dbl    |
|blocks                                        |   dbl    |
|blocks_received                               |   dbl    |
|personal_fouls                                |   dbl    |
|fouls_drawn                                   |   dbl    |
|total_fouls                                   |   dbl    |
|bonus_fouls                                   |   dbl    |
|points_in_the_paint                           |   dbl    |
|second_chance_points                          |   dbl    |
|points_from_turnovers                         |   dbl    |
|bench_points                                  |   dbl    |
|fast_break_points                             |   dbl    |
|team_index_rating                             |   dbl    |
|team_index_rating_2                           |   dbl    |
|team_index_rating_3                           |   dbl    |
|team_index_rating_4                           |   dbl    |
|team_index_rating_5                           |   dbl    |
|team_index_rating_6                           |   dbl    |
|team_index_rating_7                           |   dbl    |
|team_fouls                                    |   dbl    |
|team_turnovers                                |   dbl    |
|team_rebounds                                 |   dbl    |
|team_defensive_rebounds                       |   dbl    |
|team_offensive_rebounds                       |   dbl    |
|period_1_score                                |   dbl    |
|period_2_score                                |   dbl    |
|period_3_score                                |   dbl    |
|period_4_score                                |   dbl    |
|biggest_lead                                  |   dbl    |
|biggest_scoring_run                           |   dbl    |
|time_leading                                  |   dbl    |
|lead_changes                                  |   dbl    |
|times_scores_level                            |   dbl    |
|timeouts_left                                 |   dbl    |
|head_coach                                    |   chr    |
|assistant_coach_1                             |   chr    |
|assistant_coach_2                             |   chr    |
|international_team_name                       |   chr    |
|international_short_name                      |   chr    |
|international_code                            |   chr    |
|logo                                          |   chr    |
|logo_t_url                                    |   chr    |
|logo_t_size                                   |   chr    |
|logo_t_height                                 |   dbl    |
|logo_t_width                                  |   dbl    |
|logo_t_bytes                                  |   dbl    |
|logo_s_url                                    |   chr    |
|logo_s_size                                   |   chr    |
|logo_s_height                                 |   dbl    |
|logo_s_width                                  |   dbl    |
|logo_s_bytes                                  |   dbl    |
==============================================   ===========

}
